Add editing option in Item page

+++ Karma/Services/JsonFilePostService.cs
@@ -5,6 +5,8 @@ using System.Text.Json;
 using Karma.Models;
 using Microsoft.AspNetCore.Hosting;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
+using System;
 
 namespace Karma.Services
 {
@@ -41,26 +43,98 @@ namespace Karma.Services
                 new JsonSerializerOptions {WriteIndented = true}));
         }
 
-        public T GetPost(JsonFilePostService<T> postService, string id)
+        public T GetPost(string id)
         {
-            IEnumerable<T> posts = postService.GetPosts();
+            IEnumerable<T> posts = GetPosts();
             T post = posts.FirstOrDefault<T>(post => post.ID == id);
 
             return post;
         }
 
-        public void DeletePost(JsonFilePostService<ItemPost> submitService, string id)
+        public void DeletePost(JsonFilePostService<ItemPost> submitService, string id, bool deletePicture = true)
         {
             IEnumerable<ItemPost> posts = submitService.GetPosts();
             ItemPost post = posts.FirstOrDefault<ItemPost>(post => post.ID == id);
 
-            if(post.Picture != "noimage.jpg")
+            if (deletePicture)
             {
-                string filePath = Path.Combine(WebHostEnvironment.WebRootPath, "images", post.Picture);
-                System.IO.File.Delete(filePath);
+                if (post.Picture != "noimage.jpg")
+                {
+                    string filePath = Path.Combine(WebHostEnvironment.WebRootPath, "images", post.Picture);
+                    System.IO.File.Delete(filePath);
+                }
+
             }
 
             submitService.RefreshPosts(posts.Where(post => post.ID != id));
+        }
+
+        public void UpdatePost(JsonFilePostService<ItemPost> submitService, ItemPost newItem, string id, IFormFile newPhoto = null)
+        {
+            IEnumerable<ItemPost> items = submitService.GetPosts();
+            ItemPost item = submitService.GetPost(id);
+
+            bool deletePicFlag = false;
+
+            if(newItem.Title != null) item.Title = newItem.Title;
+            if(newItem.PhoneNumber != null) item.PhoneNumber = newItem.PhoneNumber;
+            if(newItem.PosterName  != null) item.PosterName = newItem.PosterName;
+            if(newItem.Description != null) item.Description = newItem.Description;
+            if(newPhoto            != null)
+            {
+                item.Picture = ProcessUploadedFile(newPhoto);
+                deletePicFlag = true;
+            }
+
+            DeletePost(submitService, id, deletePicFlag);
+            AddPost(submitService, item, newPhoto, false);
+        }
+
+        public void AddPost(JsonFilePostService<ItemPost> submitService, ItemPost item, IFormFile photo, bool applyDefaultPic = true)
+        {
+            if (photo != null)
+            {
+                if (item.Picture != null) //If our Item already has a picture path string, we should delete it first to upload a new one
+                {
+                    string filePath = Path.Combine(WebHostEnvironment.WebRootPath, "images", item.Picture);
+                    System.IO.File.Delete(filePath);
+                }
+
+                item.Picture = ProcessUploadedFile(photo); //Check definition
+            }
+	        else
+	        {
+                if(applyDefaultPic)
+		            item.Picture = "noimage.jpg";
+	        }
+                item.Date = DateTime.Now;
+                item.ID   = Guid.NewGuid().ToString();
+
+                IEnumerable<ItemPost> Submits = submitService.GetPosts().
+                Append<ItemPost>(item);
+
+	        Submits = Submits.OrderByDescending(item => item.State).ThenByDescending(item => item.Title);
+
+                submitService.RefreshPosts(Submits);
+
+        }
+        internal string ProcessUploadedFile(IFormFile photo)
+        {
+            string uniqueFileName = null;
+
+            if (photo != null)
+            {
+                string uploadsFolder =
+                    Path.Combine(WebHostEnvironment.WebRootPath, "images");
+                uniqueFileName = Guid.NewGuid().ToString() + "_" + photo.FileName;
+                string filePath = Path.Combine(uploadsFolder, uniqueFileName);
+                using (var fileStream = new FileStream(filePath, FileMode.Create))
+                {
+                    photo.CopyTo(fileStream);
+                }
+            }
+
+            return uniqueFileName;
         }
     }
 }
